backend/requirements.txt
Django==6.0.4
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
django-cors-headers==4.9.0
psycopg2-binary==2.9.10
python-dotenv==1.1.1
